Factor out the bulk of the react components for properties (improves fast refresh)
