+++ packages/dev/inspector-v2/src/components/properties/commonGeneralProperties.tsx
