+++ packages/dev/inspector-v2/src/components/properties/meshGeneralProperties.tsx
@@ -3,7 +3,7 @@ import type { AbstractMesh } from "core/index";
 
 import type { FunctionComponent } from "react";
 
-import { useObservableState } from "../../../../hooks/observableHooks";
+import { useObservableState } from "../../hooks/observableHooks";
 
 export const MeshGeneralProperties: FunctionComponent<{ entity: AbstractMesh }> = ({ entity: mesh }) => {
     // Use the observable to keep keep state up-to-date and re-render the component when it changes.
