+++ packages/dev/inspector-v2/src/components/properties/meshTransformProperties.tsx
