+++ packages/dev/inspector-v2/src/services/panes/properties/commonPropertiesService.ts
@@ -1,8 +1,8 @@
-import type { ServiceDefinition } from "../../../../modularity/serviceDefinition";
-import type { IPropertiesService } from "../propertiesService";
+import type { ServiceDefinition } from "../../../modularity/serviceDefinition";
+import type { IPropertiesService } from "./propertiesService";
 
-import { PropertiesServiceIdentity } from "../propertiesService";
-import { CommonGeneralProperties } from "./commonGeneralProperties";
+import { PropertiesServiceIdentity } from "./propertiesService";
+import { CommonGeneralProperties } from "../../../components/properties/commonGeneralProperties";
 
 type CommonEntity = {
     id?: number;
