+++ packages/dev/inspector-v2/src/services/panes/properties/meshPropertiesService.ts
@@ -1,13 +1,13 @@
-import type { ServiceDefinition } from "../../../../modularity/serviceDefinition";
-import type { IPropertiesService } from "../propertiesService";
+import type { ServiceDefinition } from "../../../modularity/serviceDefinition";
+import type { IPropertiesService } from "./propertiesService";
 
 import { AbstractMesh } from "core/Meshes/abstractMesh";
 
-import { GeneralPropertiesSectionIdentity } from "../common/commonPropertiesService";
-import { PropertiesServiceIdentity } from "../propertiesService";
-import { MeshAdvancedProperties } from "./meshAdvancedProperties";
-import { MeshGeneralProperties } from "./meshGeneralProperties";
-import { MeshTransformProperties } from "./meshTransformProperties";
+import { GeneralPropertiesSectionIdentity } from "./commonPropertiesService";
+import { PropertiesServiceIdentity } from "./propertiesService";
+import { MeshAdvancedProperties } from "../../../components/properties/meshAdvancedProperties";
+import { MeshGeneralProperties } from "../../../components/properties/meshGeneralProperties";
+import { MeshTransformProperties } from "../../../components/properties/meshTransformProperties";
 
 export const TransformsPropertiesSectionIdentity = Symbol("Transforms");
 export const AdvancedPropertiesSectionIdentity = Symbol("Advanced");
